requirements.txt
geopandas
rtree
descartes
tqdm
h3
plotly
googlemaps
scikit-learn
osmnx
networkx